+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Releases
   attribute.
 - Added ``thriftrw.install`` to install a Thrift file as a submodule of a
   module.
+- Expose ``thriftrw.spec.FunctionResultSpec`` with information on the return
+  and exception types of the function.
 
 
 0.1.0 (2015-08-28)